backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.5",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "python-dotenv>=1.2.2",
]
